+++ tests/unit/ClassArrayEnumType/DynamicClassArrayEnumTest.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 namespace FireMidge\Tests\ValueObject\Unit\ClassArrayEnumType;
 
 use FireMidge\Tests\ValueObject\Unit\Classes\CustomEnumType;
+use FireMidge\Tests\ValueObject\Unit\Classes\CustomEnumWithPrivateAllType;
 use FireMidge\Tests\ValueObject\Unit\Classes\DynamicClassArrayEnumType;
 use FireMidge\Tests\ValueObject\Unit\Classes\SimpleObject;
 use PHPUnit\Framework\TestCase;
@@ -19,7 +20,7 @@ class DynamicClassArrayEnumTest extends TestCase
         DynamicClassArrayEnumType::useClass(SimpleObject::class);
 
         $this->expectException(RuntimeException::class);
-        $this->expectExceptionMessage('Method FireMidge\ValueObject\IsClassArrayEnumType::all is not implemented');
+        $this->expectExceptionMessage('Method FireMidge\Tests\ValueObject\Unit\Classes\SimpleObject::all is not implemented');
 
         DynamicClassArrayEnumType::withAll();
     }
@@ -29,7 +30,17 @@ class DynamicClassArrayEnumTest extends TestCase
         DynamicClassArrayEnumType::useClass(CustomEnumType::class);
 
         $this->expectException(RuntimeException::class);
-        $this->expectExceptionMessage('Method FireMidge\ValueObject\IsClassArrayEnumType::all is not callable. Make sure it has no required parameters and it returns an array. Too few arguments to function FireMidge\Tests\ValueObject\Unit\Classes\CustomEnumType::all(), 0 passed and exactly 1 expected');
+        $this->expectExceptionMessage('Method FireMidge\Tests\ValueObject\Unit\Classes\CustomEnumType::all is not callable. Make sure it has no required parameters and it returns an array. Too few arguments to function FireMidge\Tests\ValueObject\Unit\Classes\CustomEnumType::all(), 0 passed and exactly 1 expected');
+
+        DynamicClassArrayEnumType::withAll();
+    }
+
+    public function testAllMethodIsProtected() : void
+    {
+        DynamicClassArrayEnumType::useClass(CustomEnumWithPrivateAllType::class);
+
+        $this->expectException(RuntimeException::class);
+        $this->expectExceptionMessage('Method FireMidge\Tests\ValueObject\Unit\Classes\CustomEnumWithPrivateAllType::all is not implemented or is not callable. Make sure it is public.');
 
         DynamicClassArrayEnumType::withAll();
     }