+++ tests/unit/CollectionType/StringClassCollectionTest.php
@@ -12,7 +12,11 @@ use PHPUnit\Framework\TestCase;
 use stdClass;
 
 /**
+ * Note: Read TransformIntArrayEnumTest for information on why we're saying that this test also covers BasicStringCollectionType.
+ *  It's the same reason as to why we're saying something similar there. (For Infection mutation test coverage.)
+ *
  * @covers \FireMidge\Tests\ValueObject\Unit\Classes\StringClassCollectionType
+ * @covers \FireMidge\Tests\ValueObject\Unit\Classes\BasicStringCollectionType
  * @uses \FireMidge\Tests\ValueObject\Unit\Classes\SimpleStringType
  */
 class StringClassCollectionTest extends TestCase